+++ dev_scripts/notebooks/run_notebook.py
@@ -29,6 +29,7 @@ _LOG = logging.getLogger(__name__)
 # #############################################################################
 
 
+# TODO(gp): Reuse hjupyter.run_notebook if possible
 def _run_notebook(
     config: cconfig.Config,
     notebook_file: str,
