+++ helpers/unit_test.py
@@ -1567,62 +1567,3 @@ class TestCase(unittest.TestCase):
     def _to_error(self, msg: str) -> None:
         self._error_msg += msg + "\n"
         _LOG.error(msg)
-
-
-# #############################################################################
-# Notebook testing.
-# #############################################################################
-
-
-def run_notebook(
-    file_name: str,
-    scratch_dir: str,
-    config_builder: Optional[str] = None,
-    idx: int = 0,
-) -> None:
-    """
-    Run jupyter notebook.
-
-    `core.config_builders.get_config_from_env()` supports passing in a config
-    only through a path to a config builder function that returns a list of
-    configs, and a config index from that list.
-
-    Assert if the notebook doesn't complete successfully.
-
-    :param file_name: path to the notebook to run. If this is a .py file,
-        convert to .ipynb first
-    :param scratch_dir: temporary dir storing the output
-    :param config_builder: path to config builder function that returns a list
-        of configs
-    :param idx: index of target config in the config list
-    """
-    file_name = os.path.abspath(file_name)
-    dbg.dassert_exists(file_name)
-    dbg.dassert_exists(scratch_dir)
-    # Build command line.
-    cmd = []
-    # Convert .py file into .ipynb if needed.
-    root, ext = os.path.splitext(file_name)
-    if ext == ".ipynb":
-        notebook_name = file_name
-    elif ext == ".py":
-        cmd.append(f"jupytext --update --to notebook {file_name}; ")
-        notebook_name = f"{root}.ipynb"
-    else:
-        raise ValueError(f"Unsupported file format for `file_name`='{file_name}'")
-    # Export config variables.
-    if config_builder is not None:
-        cmd.append(f'export __CONFIG_BUILDER__="{config_builder}"; ')
-        cmd.append(f'export __CONFIG_IDX__="{idx}"; ')
-        cmd.append(f'export __CONFIG_DST_DIR__="{scratch_dir}" ;')
-    # Execute notebook.
-    cmd.append("cd %s && " % scratch_dir)
-    cmd.append("jupyter nbconvert %s" % notebook_name)
-    cmd.append("--execute")
-    cmd.append("--to html")
-    cmd.append("--ExecutePreprocessor.kernel_name=python")
-    # No time-out.
-    cmd.append("--ExecutePreprocessor.timeout=-1")
-    # Execute.
-    cmd_as_str = " ".join(cmd)
-    hsinte.system(cmd_as_str, abort_on_error=True)
